Use generic code for saving auth token
