+++ pkg/util/filesystem/permission_nowindows.go
@@ -6,7 +6,12 @@
 
 package filesystem
 
-import "os"
+import (
+	"fmt"
+	"os"
+	"os/user"
+	"strconv"
+)
 
 // Permission handles permissions for Unix and Windows
 type Permission struct{}
@@ -18,5 +23,22 @@ func NewPermission() (*Permission, error) {
 
 // RestrictAccessToUser restricts the access to the user (chmod 700)
 func (p *Permission) RestrictAccessToUser(path string) error {
-	return os.Chmod(path, 0700)
+	usr, err := user.Lookup("dd-agent")
+	if err == nil {
+		usrID, err := strconv.Atoi(usr.Uid)
+		if err != nil {
+			return fmt.Errorf("couldn't parse UID (%s): %w", usr.Uid, err)
+		}
+
+		grpID, err := strconv.Atoi(usr.Gid)
+		if err != nil {
+			return fmt.Errorf("couldn't parse GID (%s): %w", usr.Gid, err)
+		}
+
+		if err = os.Chown(path, usrID, grpID); err != nil {
+			return fmt.Errorf("couldn't set user and group owner for %s: %w", path, err)
+		}
+	}
+
+	return nil
 }
